Rename apt sources file to fix fish installation on Debian
With the old file name, apt wasn't recognizing the repository used
to install the right fish version, and when we tried to install fish,
we were installing the outdated version from the Debian repositories.

By changing the file name of the apt sources file, apt is now able to
find the fish repository where we can get the latest fish version.